Refactored registry opa tests.
